Rename to LoadableCardButton and switch dislike icon
